+++ src/app/admin/components/index.ts
@@ -1,0 +1,16 @@
+import { AdminComponent } from './admin';
+import { ProductsComponent } from './products';
+import { ProductFormComponent } from './product-form';
+import { OrdersComponent } from './orders';
+
+export * from './products';
+export * from './admin';
+export * from './product-form';
+export * from './orders';
+
+export const components = [
+  AdminComponent,
+  ProductsComponent,
+  ProductFormComponent,
+  OrdersComponent
+];