+++ src/app/admin/components/products/products.component.ts
@@ -1,0 +1,21 @@
+import { Component, OnInit } from '@angular/core';
+import { Observable } from 'rxjs';
+
+import { ProductModel, ProductsService } from '../../../products';
+
+@Component({
+  selector: 'sh-products',
+  templateUrl: './products.component.html',
+  styleUrls: ['./products.component.scss']
+})
+export class ProductsComponent implements OnInit {
+  products$: Observable<ProductModel[]> = this.productsService.getProducts$();
+
+  constructor(
+    private productsService: ProductsService
+  ) { }
+
+  ngOnInit() {
+  }
+
+}