+++ src/app/auth/auth-routing.module.ts
@@ -1,0 +1,21 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+
+import { LoginComponent } from './components/login';
+
+const routes: Routes = [{
+  path: '',
+  pathMatch: 'full',
+  redirectTo: 'login'
+}, {
+  path: 'login',
+  component: LoginComponent
+}];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule]
+})
+export class AuthRoutingModule {
+  static components = [ LoginComponent ];
+}