+++ src/app/auth/models/user.model.ts
@@ -1,0 +1,5 @@
+export interface UserModel {
+  id: number;
+  email: string;
+  login: string;
+}