+++ src/app/cart/cart.module.ts
@@ -4,6 +4,7 @@ import { CommonModule } from '@angular/common';
 import { CartComponent, CartItemComponent, CartListComponent } from './components';
 
 import { SharedModule } from '../shared';
+import { RouterModule } from '@angular/router';
 
 @NgModule({
   declarations: [
@@ -17,7 +18,8 @@ import { SharedModule } from '../shared';
   imports: [
     CommonModule,
 
-    SharedModule
+    SharedModule,
+    RouterModule
   ]
 })
 export class CartModule { }