+++ src/app/cart/index.ts
@@ -1,1 +1,2 @@
 export { CartModule } from './cart.module';
+export * from './components';