+++ src/app/components/index.ts
@@ -1,2 +1,13 @@
+import { FirstComponent } from './first';
+import { AboutComponent } from './about';
+import { PageNotFoundComponent } from './page-not-found';
+
+export const components = [
+  FirstComponent,
+  AboutComponent,
+  PageNotFoundComponent
+];
+
 export * from './first';
 export * from './about';
+export * from './page-not-found';