+++ src/app/orders/components/index.ts
@@ -1,0 +1,7 @@
+import { CreateOrderComponent } from './create-order';
+
+export const components = [
+  CreateOrderComponent
+];
+
+export * from './create-order';