+++ src/app/orders/orders.module.ts
@@ -1,10 +1,11 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-
+import { components } from './components';
 
 @NgModule({
-  declarations: [],
+  declarations: [...components],
+  exports: [...components],
   imports: [
     CommonModule
   ]