+++ src/app/products/products.module.ts
@@ -2,6 +2,8 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { ProductComponent, ProductListComponent } from './components';
+import { RouterModule } from '@angular/router';
+import { ProductDetailsComponent } from './components/product-details/product-details.component';
 
 const components = [
   ProductComponent,
@@ -9,8 +11,8 @@ const components = [
 ];
 
 @NgModule({
-  declarations: [...components],
-  imports: [CommonModule],
+  declarations: [...components, ProductDetailsComponent],
+  imports: [CommonModule, RouterModule],
   exports: [ProductListComponent]
 })
 export class ProductsModule {}