+++ src/app/shared/components/index.ts
@@ -1,5 +1,13 @@
 import { SelectComponent } from './select';
+import { NavComponent } from './nav';
+import { CartIndicatorComponent } from './cart-indicator';
+
+export const exportComponents = [
+  SelectComponent,
+  NavComponent
+];
 
 export const components = [
-  SelectComponent
+  ...exportComponents,
+  CartIndicatorComponent
 ];