+++ src/app/shared/directives/index.ts
@@ -1,7 +1,9 @@
 import { HighlightDirective } from './highlight.directive';
 import { PrettyDirective } from './pretty.directive';
+import { CanViewDirective } from './can-view.directive';
 
 export const directives = [
   HighlightDirective,
-  PrettyDirective
+  PrettyDirective,
+  CanViewDirective
 ];